requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
redis==5.0.1
redis-server
rq==1.15.1
watchdog==4.0.0